Refactored kepware library
