--- interfaces/kepware2/index.js
@@ -1,127 +1,0 @@
-/**
- * Copyright (c) 2018 PTC
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-
-const KepwareClient = require('./kepwareClient');
-var server = require('../../../../libraries/hardwareInterfaces');
-var settings = server.loadHardwareInterface(__dirname);
-var kepwareClient = new KepwareClient('Spatial Toolbox');
-
-exports.enabled = settings('enabled', false);
-exports.configurable = true; // can be turned on/off/adjusted from the web frontend
-
-const parseId = (nodeId) => {
-  const split = nodeId.split('.');
-  if (split.length === 3) {
-    return {
-      channelName: split[0],
-      objectName: split[1],
-      frameName: split[2], // If not in a group, consider it in a group of itself
-      tagName: split[2]
-    }
-  }
-  return {
-    channelName: split[0],
-    objectName: split[1],
-    frameName: split.slice(2, split.length - 1).join('_'),
-    tagName: split[split.length - 1]
-  }
-}
-
-const createNode = (tag) => {
-  const {objectName, frameName, tagName} = parseId(tag.nodeId);
-  kepwareClient.getTagPermissions(tag).then(permissions => {
-    if (permissions.canRead) {
-      server.addNode(`${settings('name')}_${objectName}`, frameName, `read${tagName}`, 'node', {x:50,y:50});
-    }
-    if (permissions.canWrite) {
-      server.addNode(`${settings('name')}_${objectName}`, frameName, `write${tagName}`, 'node', {x:-50,y:50});
-      server.addReadListener(`${settings('name')}_${objectName}`, frameName, `write${tagName}`, data => {
-        kepwareClient.writeTag(tag, data.value);
-      });
-    }
-  });
-  kepwareClient.monitorTag(tag, value => {
-    server.write(`${settings('name')}_${objectName}`, frameName, `read${tagName}`, value);
-  })
-}
-
-const removeNode = (tag) => {
-  const {objectName, frameName, tagName} = parseId(tag.nodeId);
-  server.removeNode(`${settings('name')}_${objectName}`, frameName, `read${tagName}`);
-  server.removeNode(`${settings('name')}_${objectName}`, frameName, `write${tagName}`);
-}
-
-const onKepwareConnect = () => {
-  kepwareClient.getAllTags().then(allTags => {
-    const enabledTags = settings('enabledTags', []);
-    console.log(`Discovered ${allTags.length} tags on OPC UA endpoint`);
-    allTags.forEach(tag => {
-      if (enabledTags.find(enabledTag => enabledTag.nodeId === tag.nodeId)) {
-        createNode(tag);
-      } else {
-        removeNode(tag);
-      }
-    })
-    exports.allTags = allTags;
-  });
-}
-
-if (exports.enabled) {
-  server.addEventListener('reset', function() {
-    console.log('Resetting Kepware interface');
-    exports.allTags = [];
-    kepwareClient.disconnect();
-    setup();
-  });
-
-  function setup() { // eslint-disable-line no-inner-declarations
-    settings = server.loadHardwareInterface(__dirname);
-    
-    /**
-    * These settings will be exposed to the webFrontend to potentially be modified
-    */
-    exports.settings = {
-      endpointUrl: {
-        value: settings('endpointUrl', ''), // "opc.tcp://host:49330"
-        type: 'text',
-        helpText: 'The url and port (default: 49330) of the Kepware OPC UA endpoint you want to connect to.'
-      },
-      username: {
-        value: settings('username', ''),
-        type: 'text',
-        helpText: 'The username used to connect to the Kepware OPC UA server'
-      },
-      password: {
-        value: settings('password', ''),
-        type: 'text',
-        helpText: 'The password used to connect to the Kepware OPC UA server'
-      },
-      name: {
-        value: settings('name', 'kepware'),
-        type: 'text',
-        helpText: 'The prefix of the Reality Object where nodes for each tag will be created.'
-      },
-      enabledTags: {
-        value: settings('enabledTags', []),
-        type: 'array',
-        helpText: 'The tags for which nodes will be created.',
-        hidden: true
-      }
-    };
-    
-    if (settings('enabled', false)) {
-      const url = settings('endpointUrl', '');
-      const credentials = {
-        userName: settings('username', ''),
-        password: settings('password', '')
-      }
-      kepwareClient.connect(url, credentials).then(onKepwareConnect);
-      server.enableDeveloperUI(true);
-    }
-  }
-}
